refactor: move execution metadata into execution specs
