+++ polylogue/scenarios/__init__.py
@@ -18,7 +18,6 @@ from .execution import (
     ExecutionSpec,
     PipelineProbeInputMode,
     PipelineProbeRequest,
-    command_execution,
     composite_execution,
     devtools_execution,
     memory_budget_execution,
@@ -55,7 +54,6 @@ __all__ = [
     "build_default_corpus_specs",
     "build_corpus_scenarios",
     "build_inferred_corpus_specs",
-    "command_execution",
     "composite_execution",
     "devtools_execution",
     "flatten_corpus_specs",
